# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: 供应链金融系统 API文档
  description: 基于FISCO BCOS区块链的供应链金融系统接口文档
  contact:
    name: 开发团队
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Inferred Url
...
paths:
...
  /api/account:
    get:
      tags:
      - bcos-controller
      summary: getAccountInfo
      operationId: getAccountInfoUsingGET
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/block:
    get:
      tags:
      - bcos-controller
      summary: getBlock
      operationId: getBlockUsingGET
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/block/latest:
    get:
      tags:
      - bcos-controller
      summary: getLatestBlockNumber
      operationId: getLatestBlockNumberUsingGET
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/block/{number}:
    get:
      tags:
      - bcos-controller
      summary: getBlockByNumber
      operationId: getBlockByNumberUsingGET
      parameters:
      - name: number
        in: path
        description: number
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/contract/deploy:
    post:
      tags:
      - bcos-controller
      summary: deployHelloWorld
      operationId: deployHelloWorldUsingPOST
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/contract/{address}/get:
    get:
      tags:
      - bcos-controller
      summary: callHelloWorldGet
      operationId: callHelloWorldGetUsingGET
      parameters:
      - name: address
        in: path
        description: address
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
  /api/contract/{address}/set:
    post:
      tags:
      - bcos-controller
      summary: callHelloWorldSet
      operationId: callHelloWorldSetUsingPOST
      parameters:
      - name: address
        in: path
        description: address
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
...
  /api/nodes:
    get:
      tags:
      - bcos-controller
      summary: getPeers
      operationId: getPeersUsingGET
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Authorization:
        - global
...
components:
...
  securitySchemes:
    Authorization:
      type: apiKey
      name: Authorization
      in: header